emacs: fixed colors

+++ restore_dotfiles.sh
@@ -4,6 +4,7 @@ DARWIN_PKGS=('emacs' 'vim' 'zsh' 'tmux' 'starship' 'wget' 'radare2')
 
 ! command -v stow 1>/dev/null && echo '[!] Please install stow!' && exit 1
 ! command -v curl 1>/dev/null && echo '[!] Please install curl!' && exit 1
+! command -v git  1>/dev/null && echo '[!] Please install git!' && exit 2
 
 echo "[-] Start stowing"
 if [[ `uname -a | cut -d' ' -f 1` == 'Darwin' ]]; then
@@ -11,7 +12,8 @@ if [[ `uname -a | cut -d' ' -f 1` == 'Darwin' ]]; then
 		stow -v 1 -R -t ~ $PKG 2>&1 | grep 'WARNING! unstowing' -A1
 	done
 else
-	find * -maxdepth 0 -type d -exec stow -v 1 -R -t ~ {} \; 2>&1 | grep 'WARNING! unstowing' -A1
+	find * -maxdepth 0 -type d | grep -v '^common$' | \
+		xargs -I@ stow -v 1 -R -t ~ @ 2>&1 | grep 'WARNING! unstowing' -A1
 fi
 echo "[+] Done stowing"
 
@@ -50,7 +52,9 @@ echo "[+] Done setting XDG dirs"
 echo "[-] Setting up vim"
 
 # Install spell
-curl -L 'http://ftp.vim.org/pub/vim/runtime/spell/it.utf-8.spl' -o "$XDG_DATA_HOME/vim/spell/it.utf-8.spl"
+if [[ ! -f "$XDG_DATA_HOME/vim/spell/it.utf-8.spl" ]]; then
+	curl -L 'http://ftp.vim.org/pub/vim/runtime/spell/it.utf-8.spl' -o "$XDG_DATA_HOME/vim/spell/it.utf-8.spl"
+fi
 
 echo '[+] Done setting up vim'
 
@@ -59,5 +63,9 @@ if [[ ! -d "${XDG_CONFIG_HOME}/tmux/plugins/tpm" ]]; then
 	git clone https://github.com/tmux-plugins/tpm "${XDG_CONFIG_HOME}/tmux/plugins/tpm"
 fi
 echo '[+] Done setting up tmux'
+
+echo '[-] Do common things'
+
+find ./common -type f -exec {} \;
 
 echo '[+] All done!'
